+++ modules/App/views/auth/login.php
@@ -20,7 +20,7 @@
                 </div>
 
                 <div class="kiss-margin">
-                    <button class="kiss-button kiss-button-primary kiss-width-1-1"><?=t('Login')?></button>
+                    <button class="kiss-button kiss-button-primary kiss-width-1-1"><?=t('Log in')?></button>
                 </div>
 
             </form>