+++ modules/System/views/api/index.php
@@ -65,11 +65,15 @@
                     <li v-for="(key, idx) in keys">
                         <div class="kiss-margin kiss-flex kiss-flex-middle">
                             <div class="kiss-margin-small-right"><icon>vpn_key</icon></div>
-                            <div class="kiss-size-5 kiss-flex-1 kiss-position-relative">
+                            <div class="kiss-size-5 kiss-position-relative">
                                 <strong>{{key.name}}</strong>
                                 <a class="kiss-cover" :href="$route('/system/api/key/'+key._id)"></a>
                             </div>
-                            <div class="kiss-margin-left kiss-align-right kiss-text-monospace kiss-color-muted">*****{{ key.key.substr(-5, 5)}}</div>
+                            <div class="kiss-flex-1 kiss-margin-left kiss-align-right kiss-text-monospace kiss-color-muted">*****{{ key.key.substr(-5, 5)}}</div>
+                            <div class="kiss-margin-left kiss-button-group" :class="{'kiss-disabled': !key.key}">
+                                <a class="kiss-button kiss-button-small kiss-width-1-1" href="#" @click="showApiViewer(key.key)">REST</a>
+                                <a class="kiss-button kiss-button-small kiss-width-1-1" href="#" @click="showGraphQLViewer(key.key)">GraphQL</a>
+                            </div>
                             <a class="kiss-margin-left" href="#" @click="copyApiKey(key.key)"><icon>content_copy</icon></a>
                             <a class="kiss-margin-small-left kiss-color-danger" @click="remove(key)"><icon>delete</icon></a>
                         </div>
@@ -142,12 +146,17 @@
                         });
                     },
 
-                    showApiViewer() {
-                        VueView.ui.offcanvas('system:assets/dialogs/api-viewer.js', {openApiUrl: this.$route('/system/api/openapi')}, {}, {flip: true, size: 'xxlarge'})
+                    showApiViewer(apiKey = null) {
+                        VueView.ui.offcanvas('system:assets/dialogs/api-viewer.js', {
+                            openApiUrl: this.$route('/system/api/openapi'),
+                            apiKey
+                        }, {}, {flip: true, size: 'xxlarge'})
                     },
 
-                    showGraphQLViewer() {
-                        VueView.ui.offcanvas('system:assets/dialogs/graphql-viewer.js', {}, {}, {flip: true, size: 'xxlarge'})
+                    showGraphQLViewer(apiKey = null) {
+                        VueView.ui.offcanvas('system:assets/dialogs/graphql-viewer.js', {
+                            apiKey
+                        }, {}, {flip: true, size: 'xxlarge'})
                     }
                 }
             }