+++ modules/System/views/api/rest-api-viewer.php
@@ -11,6 +11,12 @@
         render-style="read"
         load-fonts="false"
 
+        <?php if($apiKey): ?>
+        api-key-name = "api-key"
+        api-key-location = "header"
+        api-key-value = "<?=$apiKey?>"
+        <?php endif ?>
+
         bg-color="<?=($bgColor ? $bgColor : '#10131a')?>"
         text-color="<?=($textColor ? $textColor : '#fafafa')?>"
         primary-color="<?=($primaryColor ? $primaryColor : '#0e8fff')?>"