+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## WIP
 
 - Fix select boxes in dark theme (chrome on windows) #8
+- Change login (noun) to log in (verb)
 
 ## 2.0.1 (2022-07-17)