+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix select boxes in dark theme (chrome on windows) #8
 - Change login (noun) to log in (verb)
+- Fix required field check on nested repeatable items
 
 ## 2.0.1 (2022-07-17)