+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix select boxes in dark theme (chrome on windows) #8
 - Change login (noun) to log in (verb)
 - Fix required field check on nested repeatable items
+- Enable picking item by clicking on empty placeholder area (Asset + Content Link field)
 
 ## 2.0.1 (2022-07-17)