+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Change login (noun) to log in (verb)
 - Fix required field check on nested repeatable items
 - Enable picking item by clicking on empty placeholder area (Asset + Content Link field)
+- Open api playground with preselected api key
 
 ## 2.0.1 (2022-07-17)