+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix required field check on nested repeatable items
 - Enable picking item by clicking on empty placeholder area (Asset + Content Link field)
 - Open api playground with preselected api key
+- Enable clear field function also in nested fields
 
 ## 2.0.1 (2022-07-17)